significantly reworked the short term archiving
